requirements.txt
playwright
pandas
numpy
Pillow
matplotlib
scipy
